+++ tests/testapp/models.py
@@ -6,6 +6,7 @@ Created on May 7, 2011
 """
 from decimal import Decimal
 
+from django.core.validators import MinValueValidator
 from django.db import models
 
 from djmoney.models.fields import MoneyField
@@ -147,3 +148,7 @@ class DateTimeModel(models.Model):
 class ModelIssue300(models.Model):
     money = models.ForeignKey(DateTimeModel, on_delete=models.CASCADE)
     price = MoneyField(max_digits=10, decimal_places=2, default_currency='EUR', default=Decimal('0.0'))
+
+
+class ModelWithValidation(models.Model):
+    balance = MoneyField(max_digits=10, decimal_places=2, validators=[MinValueValidator(Money(100, 'GBP'))])